[Stack Monitoring] Refactor common code used by parity testing scripts (#347)
* Adding module with common utility functions used by various docs_compare scripts

* Using common code in beat/doc_compare.py

* Using common code in elasticsearch/docs_compare.py

* Using common code in kibana/docs_compare.py

* Using common code in logstash/docs_compare.py

* Set PYTHONPATH so common code can be found
